box: fix crash while nil-uuid subscribe request
Before this patch a master node could crash when it processed an
anonymous iproto subscribe request which had a nil instance uuid.
The reason of this error is a lack of appropriate check in
`box_process_subscribe` that `instance_uuid` of request is not nil.
Since `replica_by_uuid` can't find a replica with given `instance_uuid`,
the replica is being NULL and then we try to dereference it.

Now, we raise an ER_PROTOCOL error if we get a subscribe request with
nil `instance_uuid`.

Closes #11531

NO_DOC=bugfix