limbo: drop txn_limbo_pop and rename another func
txn_limbo_pop() was used a single place. And can only be used
there, because popping from the end can only mean a rollback.
Hence it can only be done by txn_limbo_abort().

Lets just inline it there.

txn_limbo_remove() was renamed to txn_limbo_pop_first() to
emphasize that the entry to remove must be the first one. Which
means it must be confirmed, but that can't be mentioned
explicitly, because is also used for async txns when don't need a
confirm (when have no sync txns in front of them in the limbo).

This should simplify the further commits a bit.

Part of #11180

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring