limbo: drop txn_limbo_pop and rename another func
txn_limbo_pop() was used a single place. And can only be used
there, because popping from the end can only mean a rollback.
Hence it can only be done by txn_limbo_abort().

Lets just inline it there.

txn_limbo_remove() was renamed to txn_limbo_pop_first() to
emphasize that the entry to remove must be the first one. Which
means it must be confirmed, but that can't be mentioned
explicitly, because is also used for async txns when don't need a
confirm (when have no sync txns in front of them in the limbo).

This should simplify the further commits a bit.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

Part of #11180
Part of #11790

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring

(cherry picked from commit 137cd00e3a97ec97f0c6dc025637827c0c839fcd)