limbo: move a couple of functions
The moved function was declared below the place where it will be
needed in the next commits. Lets move it up.

See how the commit moves only one function. This is because in
2.11 one of the moved functions doesn't exist. But the commit
title is kept from the original commit.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

Part of #11180
Part of #11790

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring

(cherry picked from commit 56886db8298974e05db69f0687b00b8115af0eff)