limbo: extract entry completion code into funcs
The code of txn_limbo_complete_fail() was repeated a couple of
times. Lets create this function to reuse it. The next commits
will need to use it more.

For consistency there is also added txn_limbo_complete_success(),
even though used in just a single place.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

Part of #11180
Part of #11790

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring

(cherry picked from commit d6a57627602b2314df30baca533f8528ca367703)