memtx: build indexes sequentially
The MemTX engine used to build primary keys of all spaces after the
initial recovery is complete. The PK build was started right on the
first index creation. This complicates the following O(n) SK sorting
algorithm and makes it consume more memory (it has to hold the tuple
mappings for all PKs in memory before we start building SKs).

Let's start building the space PK on the first insertion into it and
build PK and all SKs (if required) on the first insertion into next
space in the snapshot or at the end of the initial recovery (on the
beginning of final recovery, actually). This way we're going to have
PKs and SKs built sequentially for each next space.

NO_DOC=internal
NO_TEST=covered by existing tests
NO_CHANGELOG=internal