config: add config.etcd.http.request.interface
This option enables the `http.client`'s `interface` option for the etcd
client that is used in the etcd configuration source. It allows to
specify certain network interface for outgoing connections.

Part of tarantool/tarantool-ee#1322

NO_DOC=in the EE patch
NO_CHANGELOG=same
NO_TEST=requires several network interfaces, so can't be written to run
        everywhere