box: introduce index aggregates option
It will be used by MemCS to specify aggregates for its indexes. These
aggregates will be used for data-skipping on index scan. Such approach
is also known as BRIN-index or skip-index.

Part of tarantool/tarantool-ee#1029

NO_CHANGELOG=does not work with CE engines
NO_DOC=will be done in EE