txn: set TXN_STMT_ROLLBACK flag before calling rollback triggers
We have a flag `TXN_STMT_ROLLBACK` which indicates that the statement
is being rolled back at the moment. It is set before each call of
`engine_rollback_statement` and is cleared right after. Let's set this
flag before calling system `stmt->on_rollback` triggers - it will allow
us to recognize rollback from the functions that are called by those
triggers. Note that `TXN_IS_ROLLED_BACK` doesn't work here because it is
not set on rollback to savepoint.

The actual reason we need this flag is that it's impossible to
understand whether `index_update_def` method is called on index
modification or on its rollback, but we need to distinguish between
these two cases in order to implement filters for MemCS.

Part of tarantool/tarantool-ee#1029

NO_TEST=cannot be tested
NO_CHANGELOG=internal
NO_DOC=internal