test: rewrite box/alter-primary-index-tuple-leak-long.test.lua
The test takes too long, especially on ASAN-enabled builds - for
example, 368 seconds on our debug-asan-clang workflow. Also, the test
doesn't check anything when ASAN is disabled because it sets memtx
memory to 32MB and tries to reach this limit, but we started to silently
increase it to 64MB recently (de0731bd316706c4ff509f6302325a1d3c8736bc).

The test can be rewritten in a way simpler manner - the new test
just does a few replaces and checks the statistics. The new version of
the test is much faster - it usually gets finished in a second.

NO_DOC=test fix
NO_CHANGELOG=test fix