perf: `1mops_write` - each fiber works with its own key interval
It is open to debate whether there is a need for each fiber to work with
its own key interval. However, until now the benchmark tried to achieve
this, but contained a bug in this place. This commit fixes this bug.

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=internal