box: do not normalize `covers` twice on index alter
Currently, we normalize option `covers` on each alter, even if it
wasn't changed and this option is already normalized. Since
normalization is not idempotent (it shifts all numeric fields from
1-indexation to 0-indexation), such alter can break `covers` option.
Let's normalize `covers` only if it was updated.

Part of tarantool/tarantool-ee#1335

NO_TEST=in EE
NO_CHANGELOG=in EE
NO_DOC=in EE