sql: fix memleak in `mem_destroy()`
This patch fixes a memory leak in `mem_destroy()`. Before this patch,
`mem_set_str0_allocated()` set the size of the string to be `strlen()`,
which is incorrect - the size is actually `strlen() + 1`. Later, MEM was
freed if the size was greater than 0. This led to cases where empty
strings were not freed properly.

Part of #10532

NO_DOC=fix memleaks
NO_TEST=fix memleaks
NO_CHANGELOG=fix memleaks