sql: fix memleak in `parser_space_delete()`
This patch fixes a memory leak in `parser_space_delete()`. Before this
patch, if a space with the same name as the provided space already
existed, the request was considered to be `ALTER TABLE ADD COLUMN`.
In this case, there is no need to free some index definitions, because
these belong to an existing space. However, this logic is not always
correct, because if the request was `CREATE TABLE` of a table with
the name of an already existing space, the condition was true and some
temporary index definitions were not freed.

Part of #10532

NO_DOC=fix memleaks
NO_TEST=fix memleaks
NO_CHANGELOG=fix memleaks