sql: fix leaks of ephemeral tables
Before this patch, there were cases where an ephemeral table was created
but assigned to a cursor more than one opcode later. This could lead to
a case where the ephemeral space was lost due to an error occurring in
one of the opcodes between creation and assignment. This patch moves
the assignment opcode to be immediately after ephemeral space creation.

Note that this patch is not a complete fix. A better solution would be,
for example, to force all opcodes that accept spaces to accept a cursor
instead, and assign a cursor in the same opcode when the space is opened
or the ephemeral space is created.

Closes #10532

NO_DOC=fix memleaks
NO_TEST=fix memleaks
NO_CHANGELOG=fix memleaks