limbo: privatize last/first entry getters
The following commits will have to create special entries in the
limbo queue which should not be seen by the outside code.

Lets hide the API not to expose the limbo entries directly, so the
code from other subsystems wouldn't be able to see those special
entries.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

Part of #11180
Part of #11790

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring

(cherry picked from commit 641991e077c2383a7f43ec816cb46b9b3882011e)