journal: rename the completion callback type
It was journal_write_async_f but it was called when the write is
already finished. A completion callback. It wasn't doing any
write. And it wasn't async itself either.

It is renamed to journal_on_write_f. This was noticed while doing
a next commit to introduce a new callback for the journal.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

In scope of #11180
In scope of #11790

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring

(cherry picked from commit 0f8d0a15d7d24f90df749f34083a2cf7ac739ac8)