journal: rename the completion callback type
It was journal_write_async_f but it was called when the write is
already finished. A completion callback. It wasn't doing any
write. And it wasn't async itself either.

It is renamed to journal_on_write_f. This was noticed while doing
a next commit to introduce a new callback for the journal.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring