journal: alias the write callback
The function type was inlined in a couple of places. Lets alias
is. It will look better in the next commit which is going to
introduce a new callback.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

In scope of #11180
In scope of #11790

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring

(cherry picked from commit bd1b79790c4ad0a56e8103549d8bb4b81496d09e)