journal: introduce sync callback
It does the same what wal_sync() was doing. But now the usages do
not depend on WAL directly.

This is done in order to prevent one another direct WAL dependency
in a next commit.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

In scope of #11180
In scope of #11790

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring

(cherry picked from commit 8672a36992224253d36fcc982132d858a7ef79ba)