txn: introduce txn_persist_all_prepared()
This function waits until all the prepared transactions are
persisted and returns the vclock saved at the moment of writing
the last txn.

This should encapsulate where are prepared txns located. Currently
it is mistakenly believed that if a txn is prepared, then it is
already either in the journal queue or in the WAL thread. But it
is not so. It might also be in the limbo, waiting for space in it.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

In scope of #11180
In scope of #11790

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring

(cherry picked from commit 86fb3a825084567b4e358246d5202ee69600a8ea)