[ok] txn: introduce txn_persist_all_prepared()
This function waits until all the prepared transactions are
persisted and returns the vclock saved at the moment of writing
the last txn.

This should encapsulate where are prepared txns located. Currently
it is mistakenly believed that if a txn is prepared, then it is
already either in the journal queue or in the WAL thread. But it
is not so. It might also be in the limbo, waiting for space in it.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring