sql: remove long unnecessary tests
This patch removes tests that are not needed due to the ASAN check
of the SQL code. These tests are quite long and may fail due to timeout.
Since we don't need them, it's better to delete them.

The `delete3.test.lua` test was introduced due to a bug in the SQLite3
allocator, which is irrelevant for us, especially after the introduction
of the ASAN check.

The tests `gh-3083-ephemeral-unref-tuples.test.lua` and
`gh-3332-tuple-format-leak.test.lua` were introduced due to memory leaks
that are now checked by ASAN.

The `in2.test.lua` test was introduced due to a bug in the SQLite3
b-tree, which is irrelevant for us.

Follow-up #10532

NO_DOC=test removal
NO_TEST=test removal
NO_CHANGELOG=test removal