limbo: flush it before taking the read-view vclock
When a new replica joins, it needs a checkpoint from the master +
vclock of the data in this checkpoint.

Previously it was mistakenly believed that just making a journal
sync + getting the journal vclock would be enough to make sure all
the checkpoint data has reached WAL and to gets vclock of that
data.

But it wasn't taken into account that there might be some prepared
txns in the limbo which aren't sent to the journal yet. Lets flush
the limbo before the journal sync to fix that.

Part of #11180

NO_DOC=bugfix
NO_CHANGELOG=already written