limbo: collect the checkpoint on last txn's commit
The limbo and Raft checkpoints are gathered for a read-view which
supposedly must contain all the confirmed transactions prepared at
the moment of the read-view creation. The read-view then is sent
to a new replica.

Those synchro-checkpoints must be precise. For instance, the
limbo's confirmed LSN must be exactly the LSN of the latest
synchro transaction in the read-view. If the replica would
receive a confirmation LSN newer than the read-view's, then the
replica would anyway receive CONFIRM from the master later and
would consider it conflicting. Resulting into a replication
failure and a necessity to reboostrap.

The patch makes txn_checkpoint get collected exactly when the last
synchro txn, prepared at the moment of the read-view creation,
gets confirmed. So the confirmation LSN in the limbo checkpoint
doesn't cover any newer transactions.

Part of #11180

NO_DOC=bugfix
NO_CHANGELOG=already written

(cherry picked from commit 43712634284a2b9d60186858f8b90c591149cbfd)