limbo: collect the checkpoint on last txn's commit
The limbo and Raft checkpoints are gathered for a read-view which
supposedly must contain all the confirmed transactions prepared at
the moment of the read-view creation. The read-view then is sent
to a new replica.

Those synchro-checkpoints must be precise. For instance, the
limbo's confirmed LSN must be exactly the LSN of the latest
synchro transaction in the read-view. If the replica would
receive a confirmation LSN newer than the read-view's, then the
replica would anyway receive CONFIRM from the master later and
would consider it conflicting. Resulting into a replication
failure and a necessity to reboostrap.

The patch makes txn_checkpoint get collected exactly when the last
synchro txn, prepared at the moment of the read-view creation,
gets confirmed. So the confirmation LSN in the limbo checkpoint
doesn't cover any newer transactions.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790.

Part of #11180
Part of #11790

NO_DOC=bugfix
NO_CHANGELOG=already written
NO_TEST=11180 doesn't exist here, and 11790 needs next commits

(cherry picked from commit 43712634284a2b9d60186858f8b90c591149cbfd)