txn: introduce txn_checkpoint
Many subsystems have checkpoints - WAL, each engine, limbo, Raft.
But somehow the transaction subsystem was left out. It wasn't
considered to be something having its own state.

But actually the transactions have quite a large state which
unites many others and requires them to be in exact sync: Raft,
limbo, journal.

Indeed, collecting those states separately might lead to mistakes
like having a too old journal vclock with a newer Raft/limbo
checkpoint. Or vice versa. Or any other possible wrong combination
when those checkpoints are just collected at different times and
might describe different datasets, if the instance is actively
committing new transactions during creation of those checkpoints.

The idea of the new txn_checkpoint module is to put the collection
of those interconnected states into one place. So it is isolated
from the code which doesn't care about the details, and so it
could be reused.

Note that this commit only moves some code. But the code is
broken. It is going to be fixed in the following commits.

Part of #11180

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring