wal: flush the queue before making a snapshot
The snapshot creation code goes to WAL to start the checkpoint. It
did so without taking into account that some entries might be in
the read-view of the checkpoint already, but not sent to WAL yet.

Can easily happen if they are, for example, waiting for space in
the journal queue.

Lets flush it before doing the WAL checkpoint.

Part of #11180

NO_DOC=bugfix

(cherry picked from commit 251b6d4ffb5fa14b258290a86ac7fe0f0b3336e2)