wal: flush the queue before making a snapshot
The snapshot creation code goes to WAL to start the checkpoint. It
did so without taking into account that some entries might be in
the read-view of the checkpoint already, but not sent to WAL yet.

Can easily happen if they are, for example, waiting for space in
the journal queue.

Lets flush it before doing the WAL checkpoint.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790 and parts of #11180.

Part of #11180
Part of #11790

NO_DOC=bugfix

(cherry picked from commit 251b6d4ffb5fa14b258290a86ac7fe0f0b3336e2)