limbo: flush it before making the snapshot
The snapshot creates a read-view, remembers the vclock of the last
persisted txn in it, and waits until all the txns in this
read-view are committed.

The old code would forget that just flushing the journal isn't
enough. There might be prepared and visible txns in the limbo
which didn't reach the journal yet.

This could result into the snapshot having a vclock older that its
dataset. Which could break the snapshot recovery, because it would
try to apply statements from xlogs which were already included
into the snapshot.

Lets flush the limbo so all the prepared txns of the read-view
reach the journal.

Part of #11180

NO_DOC=bugfix
NO_CHANGELOG=already written