limbo: make an illusion of CONFIRMs not batching
A single CONFIRM might cover multiple txns. In fact, this is an
important optimization which should significantly reduce load on
the journal under high throughput.

But it might look confusing for an external observer to see in an
on-commit trigger that the txn is committed and confirmed, but the
confirmation LSN is higher than LSN of this txn. Would look like
this txn's own confirmation was skipped.

It is ok when there are no observers. But there is one case when
there is an observer and the confirmation LSN being equal to the
txn's LSN matters. It happens when the instance wants to collect
the limbo checkpoint after a specific transaction is committed.

The checkpoint then must not cover any newer txns and hence the
CONFIRM of this txn can't batch with any newer txns.

It is being used in the replica snapshot uploading. The read-view
created for the replica must have a limbo checkpoint not covering
any newer txns. Or the replica would have a confirmation LSN newer
than its actual dataset.

Lets detect such "observers" by looking if there are any on-commit
triggers.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790.

Part of #11180
Part of #11790

NO_DOC=bugfix
NO_CHANGELOG=not reproducible on old versions
NO_TEST=11180 doesn't exist here, and 11790 needs next commits

(cherry picked from commit 03450e710c9930f39183be5d3c044152808fa47e)