limbo: introduce corner stones
Corner stone is a limbo entry which needs its own confirmed LSN.
It means, such an entry shouldn't be batched with newer entries.

So far that seems like this guarantee is enough to hold just in
memory. In WAL it is still ok to write one CONFIRM for multiple
entries.

This is needed so the checkpoint creation code (for a read-view
to be sent to replica) would be able to capture the limbo state
exactly when the last txn included into the checkpoint gets
committed. Without any newer txns and newer confirmations.

Part of #11180

NO_DOC=bugfix
NO_CHANGELOG=not reproducible on old versions