limbo: allow current term CONFIRMs to be nopified
Previously the limbo would treat receipt of CONFIRM for an
already confirmed LSN in the current term as an error. But
actually there seems to be no reason to do that.

Here is a usecase. A replica starts joining a master. The master
has some txns in the limbo. The relay would flush the txns to the
journal and get its vclock. This is going to be the snapshot
vclock of the replica.

Then the relay waits for the txns to get committed. CONFIRMs for
those txns are going to get written after the replica's vclock.

Next, the relay would save the current confirmed LSN and send it
+ the data to the replica.

Logically the replica looks good. It has some valid fully
confirmed snapshot and a correctly saved limbo checkpoint with the
right confirmed LSN.

But now the replica would start catching up and would download the
CONFIRM from the master which it actually already knows. It was in
the initial limbo checkpoint sent to this replica. It is sent
"again" because the replica's vclock doesn't include this CONFIRM
yet.

It seems logical that the replica must simply nopify such CONFIRM.
So lets do just that. Similar logic seems to be applicable to
ROLLBACK. Except that it was already idempotent in the current
term. And in the older terms its effect is already applied by
PROMOTE of the current term.

Now those strict filters only remain for PROMOTE/DEMOTE commands.

Closes #11180
Closes #11790

NO_DOC=bugfix
NO_CHANGELOG=already written

(cherry picked from commit e454545e53ea10f3b87c79df388ca3849b3e5bb9)