test: fix flaky trigger module test
There is a case that greps log in order to check that an error was
logged. On slow runner or when lots of tests are run in parallel,
logs may appear with delay - it will fail the test. Let's retry this
check to make the test stable.

NO_CHANGELOG=flaky test
NO_DOC=flaky test