config: set election mode from <...>.synchro_mode
The new `failover.replicasets.*.synchro_mode` option is added. Most of
the logic behind it works inside the `appoint_commit` endpoint called by
the failover coordinator. However, we also need some logic at applying a
configuration. It is added here.

If the option is `true` and the instance is not an anonymous replica,
set `box.cfg.election_mode` to `manual`, otherwise `off`. That's all.

It is a little problematic that a user is able to provide its own
`replication.election_mode` option, while we want to deduce our own
value. We need to validate that `replication.election_mode` is unset or
at least the same as the deduced value.

Part of tarantool/tarantool-ee#1369

NO_DOC=https://github.com/tarantool/doc/issues/5269
NO_CHANGELOG=added in the EE commit that fixes the issue