qsync: fix box_wait_quorum hang
Before this patch when there are uncommitted transactions in the limbo
and the `replication_synchro_quorum > replicaset.registered_count`, a
node performing the `box.ctl.promote` may become the leader but remain
in RO state. The cause of the problem is that the `box_wait_quorum`
hangs because the current node can't gather a sufficient number of acks.

This bug negatively influence on quorum-based replication because we
can't exit from this hanged state of the `box_wait_quorum`. These
actions can't help us to get back to normal state:
1) Performing `box.ctl.promote`, because we are already in the leader
   state.
2) Performing `box.ctl.demote`, because terms of the limbo and the raft
   state machine are different. The term of the limbo can't be changed
   until all transactions are committed.
3) Lowering the `replication_symchro_quorum`, because the
   `box_wait_quorum` is in the infinite yield until it get new acks and
   as a result, this function doesn't keep track some changes in
   `box.cfg`.
4) Connecting a new replica to replicaset, because our current node is
   in the RO state. It can't write into `box.space._cluster`.

To fix this issue, we decided to keep track changes of
`replication_synchro_quorum` param during transition to RW state. If
this param changes during the hang of `box_wait_quorum`, we consider
the current execution of this function obsolete. After that the
execution of the `box_wait_quorum` starts over with new quorum. This
solution allows the user to fix an incorrectly configured
`replication_synchro_quorum` param and get rid of hang while transition
to RW state.

Also we modify the `gh_7318_box_wait_limbo_acked` test by changing
expected result from fiber's join. It was done because in this patch
we don't throw ER_QUORUM_WAIT error when changing
`replication_synchro_quorum` during collecting acks.

Closes #11574

NO_DOC=bugfix