connpool: preserve deadlines between pool connects
If a connpool needs an instance for `connpool.call()` or
`connpool.filter()` it internally sets up a connection via a pool
abstraction and configures the connection with an idle deadline
timestamp. Not used connection would be closed after this deadline.
Otherwise it will be bumped.

Let's now talk about the cases when the connection to instance might
break. The original implementation works as follows: if a recently used
connection is requested but it appears to be broken the connection pool
forgets the old deadline timestamp and sets up a new connection with a
new timestamp.

Such an approach is rather OK since we usually expect to the idle
timeout to be (almost) constant. Newer requests would do a plain bump
of the instance deadline. But to make the connpool perform better in
such cases (see #10330) we need some kind of a background observer
that reconnects to the recently accessed instances appearing
unavailable. The idea here is to try to reconnect in a separate fiber
if the connection is broken. This reconnect differs from the usual
connect to instance in a way it should not affect the existing deadline.
So, let's preserve the deadline between internal reconnects to
instances. This allows us to specify 0 TTL which means just try to
reconnect if possible, do not change the idle deadline timestamp.

Such change also affects the following scenario. Previous implementation
of closing the unused connection behaves as follows.

* Config having `connpool.idle_timeout: 60` appears.
* It executes `connpool.call('box.info', {}, opts)` and sets up a new
  conn to i-001 with deadline `now + 60`.
* The connection to `i-001` breaks.
* Config having `connpool.idle_timeout: 10` appears.
* It executes `connpool.call('box.info', {}, opts)` and sets up a new
  conn to i-001 with a deadline `now + 10`.
* Wait 10 seconds... Conn i-001 is closed.

Now it behaves as follows.

* Config having `connpool.idle_timeout: 60` appears.
* It executes `connpool.call('box.info', {}, opts)` and sets up a conn
  to i-001 with deadline `now + 60`.
* The connection to `i-001` breaks.
* Config having `connpool.idle_timeout: 10` appears.
* It executes `connpool.call('box.info', {}, opts)` and sets up a new
  conn to i-001 with deadline `max(old_deadline, now + 10)`.
* Wait 10 seconds... The conn to i-001 is still alive.
* Wait 50 more seconds... The conn to i-001 is closed.

This is a really specific scenario. Both behaviors seems rather
reasonable and acceptable. Let us stick with the latter one.

Part of #10330

NO_DOC=the change does not really affect UX
NO_CHANGELOG=the original behavior has not been released yet
NO_TEST=will be tested as a part of another patch