connpool: reconnect to recent instances on error
This patch introduces logic that makes Tarantool connpool automatically
reconnect to recently accessed instances in case of errors.

The main aim is to make the connection pool methods `connpool.call()`
and `connpool.filter()` faster by making they behave as follows when
they need a connection to multiple instances.

* Run some of the mentioned connpool methods.
* Apply static configuration filters and find instance candidates.
* Check if there are recent connections to all of the candidates.
  - Yes, there are connections to all of them (both active and broken).
    In that case, try to use the active and available ones. If the
    connection is not available, we may guarantee that it has been also
    unavailable during past 3 seconds (hard-coded reconnect after
    interval). That means there is no need to wait and try to reconnect
    to it.
  - No, some instances have not been accessed yet.
    In that case, connect to all of the remaining in parallel and wait
    until the connect is established/failed.

The key improvement is that we no longer wait if some of the candidate
connections have failed and re-use the existing connections. This logic
will be added in the follow-up patch.

Part of #10330

NO_DOC=will be added later