journal: check present of prev_entry in the queue
On cascading rollback it is necessary to rollback in a reversed
order all the entries, ending with the one, which caused the
rollback.

The entries in the journal space-waiting queue are linked using
stailq. So cutting the tail of it to roll it back is not so
simple. Need to know the previous entry so as to cut all the next
ones.

The previous entry is saved before a next entry is added to the
queue. So on a rollback the previous entry is already known.

But it might happen, that by the time a rollback starts, the
previous entry is already gone. For example, it got enough space
and went to WAL, removing itself from the journal space-waiting
queue.

Then such an entry becomes not usable. It is at least for sure is
not in the list anymore. And might even be freed already.

Luckily, it is easy to check if the prev entry is valid. Lets do
that. When the prev entry is gone, it can only be possible when
the current entry became the first one. Then can just rollback
the entire queue.

Closes #11712
Part of #11180

NO_DOC=bugfix
NO_CHANGELOG=was not released

(cherry picked from commit 86fab98956523580d6d6e1adb39a893a5b839702)