connpool: add method to get conn in internal pool
This patch introduces a supplementary method in the internal pool used
within the `experimental.connpool` module. Previously, there was only
one method for getting a connection to instance that also tried to
reconnect on fail. This commit introduces a method that only gets the
existing connections and reuses it in a variety of places where it is
meant.

It, actually, affects behavior a little bit. For instance, in
`connpool.call()` the idle timeout is calculated from the earlier
point in the time when it accesses the candidates satisfying the static
requirements. It is no longer bumped when `call()` finds a suitable
candidate and checks if it satisfies dynamic requirements. Though it
does not affect the behavior a lot due to tight timings.

It also simplifies the logic of checking dynamic options. They no longer
establish a connection to instance. Instead they try to access.

Needed for #10330

NO_CHANGELOG=small behavior change, not released yet
NO_TEST=will be tested later
NO_DOC=small behavior change