connpool: speed up by using recent conn checks
This patch speeds up `connpool.call()` and `connpool.filter()` methods
by making them not wait for instances that have been unavailable
recently. Thanks to the reconnecting logic introduced in commit
6dc1b2acab02 ("connpool: reconnect to recent instances on error") we can
be sure that if connpool was not able to reconnect to an instance
recently we might continue and assume it is dead instead of waiting. A
detailed description of this idea is presented within the mentioned
patch.

Closes #10330

@TarantoolBot document
Title: connpool: add information on advanced connection management
Product: Tarantool Platform

`experimental.connpool` module also offers advanced automatic connection
management. It automatically closes unused connections and tries to
reconnect to recently used in case of an error. See also
`connpool.idle_timeout` and `connpool.reconnect_after` configuration
options.