test: bump test-run to new version
Note that the new test-run version isn't shipped with luatest - it has
to be installed from rocks. We use luatest 1.2.0.

Also, we have to update a few tests because `server.grep_log()` was
improved and the `box_cfg_log_file` environment was dropped. Now, one
doesn't have to pass a path if the server is stopped.

NO_DOC=test
NO_CHANGELOG=test