test: fix grep log usage in replication tests
Server logs are processed asynchronously. They used to be processed by
a separate process before luatest 1.1.0 and now, in luatest 1.1.0 and
newer they are processed by a fiber. This means that we should use the
retrying helper to check if a message is present in logs reliably.

While we are at it, let's also drop the log file name specification
because starting from luatest 1.1.0 it's deduced automatically, even
for a stopped server.

NO_DOC=test
NO_CHANGELOG=test