test: work around tests that make ctest hang
There are a few tests that check a lot of assertions in a loop:

 * test/app-luatest/gh_7311_malloc_info_test.lua
 * test/app-luatest/gh_8782_yaml_decode_char_code_test.lua
 * test/app-luatest/version_module_test.lua
 * test/config-luatest/jsonschema_config_test.lua

Each assertion is logged[^1]. Starting from luatest 1.1.0 logs are
processed in a fiber[^2] so if a test doesn't yield, it may fill up
the pipe[^3]. For an unknown reason, this makes ctest hang[^4]. Let's
work around this issue by yielding between loop iterations. Note that
we don't really need to check assertions in a loop in the malloc info
test - there we can replace all the assertions with just one after
the loop.

[^1]: https://github.com/tarantool/luatest/commit/d985997f1c57babdfbe6d63396e677d339ecd1eb
[^2]: https://github.com/tarantool/luatest/commit/2f16ac767f75901d233b4cca6d9fe2ead9757609
[^3]: https://github.com/tarantool/luatest/issues/418
[^4]: https://github.com/tarantool/tarantool/issues/11728

NO_DOC=test
NO_CHANGELOG=test