box: return pointer to the actual tuple in tuple_encode
A Lua helper `tuple_encode` encodes tuple on ibuf and returns its `rpos`
and `wpos` as pointers to the tuple beginning and end. If the ibuf is not
empty, the returned pointer to the tuple beginning will actually point
to the previously encoded tuple. Now it doesn't break anything, but
anyway such behavior is inconvinient, so let's fix this function and
return pointer to the actual beginning of the encoded tuple.

Along the way, rewrite `index:quantile()` in a simpler manner using the
new function behavior.

Part of #11221

NO_TEST=refactoring
NO_CHANGELOG=refactoring
NO_DOC=refactoring