update: handle invalid `ERRINJ_TUPLE_FIELD_COUNT_LIMIT` values
Currently we don't check the domain of the `ERRINJ_TUPLE_FIELD_COUNT_LIMIT`
value, which can be set to some arbitrary positive integer in fuzzing
tests. If the value falls out of the valid domain, it causes a subsequent
assertion to fail. To fix this, let's check the value's domain in the
error injection code block and return an error if it is invalid.

Closes #10033

NO_CHANGELOG=<error injection>
NO_DOC=<error injection>