test: fix box-luatest/error_subsystem_improvements
This patch is a follow-up to the commit
2120de9584be04ee7fcc1624fecbe3384c4c16e8 ("core: increase error logging
verbosity from C"). The test
`test_increased_error_string_conversion_verbosity` extended in this
commit assumes that the repository is cloned in the directory
"tarantool", so the path to the test file ends like the following:
| ...l/test/box-luatest/error_subsystem_improvements_test.lua

In the case when the last letter in the directory name is different, the
test fails. This patch chomps the last character in the directory name
to avoid failures.

Follows-up #11716

NO_DOC=test
NO_CHANGELOG=test