box: clean up recovery harness
We have `recovery_state` and also fire recovery events but the time the
state changed and the time recovery event is fired is different. This
does not look good. Also the time recovery event is fired seems to be
more sensible for example in case of FINISHED_RECOVERY state.

Let's remove this discreapency and fire the event when state is changed.

The patch also include next changes:

- RECOVERY_NOT_STARTED state is dropped as we don't really need it.
- internal on_recovery_state trigger is introduced, and space proceeds
  with its initializations steps based on events of this trigger. This
  is done due to FINISHED_RECOVERY state is set later now and trigger
  approach seems to be more robust/produces cleaner space API.
- box_set_replication_synchro_queue_max_size after recovery is finished
  but before syncing with replicas.
  This is done due to FINISHED_RECOVERY is set later now.

NO_TEST=internal
NO_CHANGELOG=internal
NO_DOC=internal