memtx: support allocation of arbitrary memory blocks
Memtx allocator implements memory management for tuples, where the main
part is the support of read views - a tuple is freed only when all read
views that can access the tuple are closed. We will need this functionality
not only for tuples, but for arbitrary blocks of memory.

This patch reworks alloc_tuple() to alloc(), which now returns a handle
for the memory block. The pointer to a tuple can be obtained by calling
memtx_block_get_tuple(), and to the data - by memtx_block_get_data(). The
difference is due to the fact that we don't want to expose tuple's header
to the user when non-tuple block is allocated.

Also the new method in_read_view() is introduced, it allows to determine
whether a block can be modified directly, or should be copied beforehand.
This wasn't needed for tuples, because they are immutable.

Closes #11744

NO_DOC=internal
NO_CHANGELOG=internal